requirements.txt
torch~=1.9.0
torchvision~=0.10.0
pytorch-lightning
future
pyyaml~=5.4.1
tensorboard
numpy~=1.19.5
tifffile~=2021.8.30
h5py~=3.1.0
scikit-image~=0.18.3
opencv-python~=4.5.3.56
elasticdeform~=0.4.9
progress
six
sklearn~=0.0

setuptools~=47.1.0
neuralnets~=1.1.9.1
tqdm~=4.62.3
scipy~=1.7.1
tensorflow~=2.6.0
matplotlib~=3.4.3
Pillow~=8.3.1
albumentations~=1.0.3
torchmetrics~=0.5.1
scikit-learn~=0.24.2